+++ src/main/java/ru/rsc/clicker_kombat/repository/CharacterRepository.java
@@ -6,7 +6,7 @@ import ru.rsc.clicker_kombat.model.domain.Character;
 import java.util.List;
 import java.util.UUID;
 
-public interface CharacterRepository extends JpaRepository<Character,Long> {
+public interface CharacterRepository extends JpaRepository<Character, Integer> {
 
     List<Character> getCharactersByPlayer_Id(UUID id);
 }
